+++ starter_project/example/get_android_dependencies.ps1
@@ -1,3 +1,3 @@
 $currentDir = (Get-Item -Path ".\" -Verbose).FullName
 [Net.ServicePointManager]::SecurityProtocol = [Net.SecurityProtocolType]::Tls12
-Invoke-WebRequest -OutFile "$currentDir\android_dependencies\com.tuarua.frekotlin-1.9.1.ane" -Uri https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/kotlin/com.tuarua.frekotlin-1.9.1.ane?raw=true
+Invoke-WebRequest -OutFile "$currentDir\android_dependencies\com.tuarua.frekotlin-1.9.5.ane" -Uri https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/kotlin/com.tuarua.frekotlin-1.9.5.ane?raw=true